+++ src/main/java/com/sb/financeq/entities/MovementDTO.java
@@ -3,9 +3,7 @@ package com.sb.financeq.entities;
 import com.sb.financeq.entities.enums.Category;
 import com.sb.financeq.entities.enums.Status;
 
-import java.math.BigDecimal;
-
 public record MovementDTO(
-        Integer movementId, String title, BigDecimal amount, Category category, Status status, String movementDate
+        Integer movementId, String title, float amount, Category category, Status status, String movementDate
 ) {
 }
